Make sim and flow use ARCUBE_INSTALL_DIR

+++ run-ndlar-flow/install_ndlar_flow.sh
@@ -4,12 +4,38 @@ set -o errexit
 
 module load python/3.11
 
+installDir=${1:-.}
+venvName=flow.venv
+
+if [[ -e "$installDir/$venvName" ]]; then
+  echo "$installDir/$venvName already exists; delete it then run me again"
+  exit 1
+fi
+
+if [[ -e "$installDir/h5flow" ]]; then
+  echo "$installDir/h5flow already exists; delete it then run me again"
+  exit 1
+fi
+
+if [[ -e "$installDir/ndlar_flow" ]]; then
+  echo "$installDir/ndlar_flow already exists; delete it then run me again"
+  exit 1
+fi
+
+staticDir=$(realpath ./static)
+
+mkdir -p "$installDir"
+cd "$installDir"
+
+python -m venv "$venvName"
+source "$venvName"/bin/activate
+
 python -m venv flow.venv
 source flow.venv/bin/activate
 pip install --upgrade pip setuptools wheel
 
 # install h5flow
-git clone https://github.com/larpix/h5flow.git
+git clone -b main https://github.com/larpix/h5flow.git
 cd h5flow
 pip install -e .
 cd ..
@@ -22,8 +48,15 @@ cd ..
 
 datadir=ndlar_flow/data/proto_nd_flow
 
-cp static/multi_tile_layout-2.4.16.yaml $datadir
-cp static/runlist-2x2-mcexample.txt $datadir
-cp static/light_module_desc-0.0.0.yaml $datadir
+cp "$staticDir"/multi_tile_layout-2.4.16.yaml $datadir
+cp "$staticDir"/runlist-2x2-mcexample.txt $datadir
+cp "$staticDir"/light_module_desc-0.0.0.yaml $datadir
 
-cp ../run-larnd-sim/larnd-sim/larndsim/detector_properties/2x2.yaml $datadir
+if [[ "$installDir" == "." ]]; then
+    cp ../run-larnd-sim/larnd-sim/larndsim/detector_properties/2x2.yaml $datadir
+elif [[ -e "$installDir/larnd-sim" ]]; then
+    cp "$installDir"/larnd-sim/larndsim/detector_properties/2x2.yaml $datadir
+else
+    echo "Couldn't find larnd-sim/larndsim/detector_properties/2x2.yaml"
+    echo "Please manually copy it into $datadir"
+fi
